requirements.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.12.0
